requirements.txt
asgiref==3.2.3
certifi==2019.11.28
Django==3.0.3
django-mysql==3.3.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
mysqlclient==1.4.6
PyJWT==1.7.1
pytz==2019.3
sqlparse==0.3.0
